# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mentor Hub mentor API
  description: |
    A Flask + MongoDB API for the mentor service.
    
    ## Authentication
    All endpoints (except `/metrics`) require JWT authentication via the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    
    For local development, obtain a JWT from the Developer Edition welcome / IdP flow, or use a
    test token from `test/e2e/e2e_auth.py` that matches your `JWT_SECRET` and expected claims.
  version: 0.1.0
  contact:
    name: Mentor Forge
    email: info@agile-learning-institute.com
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8391
    description: Local development server
...
paths:
...
  /api/Encounter:
    get:
      tags:
        - Encounter
      summary: Get all Encounter documents
      description: Retrieve infinite scroll batch of sorted, filtered Encounter documents
      operationId: getEncounters
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-Encounter
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description, status, created.at_time, saved.at_time)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved controls
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Control'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Encounter
      summary: Create a new Encounter document
      description: Create a new Encounter document (requires staff or admin role)
      operationId: createEncounter
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncounterInput'
      responses:
        '201':
          description: Encounter created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Encounter/{EncounterId}:
    get:
      tags:
        - Encounter
      summary: Get Encounter by EncounterId
      description: Retrieve a specific Encounter document by its EncounterId
      operationId: getEncounter
      security:
        - bearerAuth: []
      parameters:
        - name: EncounterId
          in: path
          required: true
          description: The Encounter ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved Encounter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Encounter'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags:
        - Encounter
      summary: Update Encounter document
      description: Update a Encounter document (requires admin role). Cannot update _id, created, or saved fields.
      operationId: updateEncounter
      security:
        - bearerAuth: []
      parameters:
        - name: EncounterId
          in: path
          required: true
          description: The Encounter ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncounterUpdate'
      responses:
        '200':
          description: Encounter updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Encounter'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Event:
    get:
      tags:
        - Event
      summary: Get all Event documents
      description: Retrieve infinite scroll batch of sorted, filtered Event documents
      operationId: getEvents
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-Event
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description, created.at_time)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved Events
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Event
      summary: Create a new Event document
      description: Create a new Event document (requires staff or admin role)
      operationId: createEvent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Event/{EventId}:
    get:
      tags:
        - Event
      summary: Get Event by EventId
      description: Retrieve a specific Event document by its EventId
      operationId: getEvent
      security:
        - bearerAuth: []
      parameters:
        - name: EventId
          in: path
          required: true
          description: The Event ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved Event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by your identity provider (Developer Edition welcome flow for local dev).
        Include the token in the Authorization header: `Bearer <token>`
...
  responses:
    BadRequest:
      description: Bad Request - Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Unauthorized
    Breadcrumb:
      type: object
      description: Audit trail information
      additionalProperties: false
      required:
        - from_ip
        - by_user
        - at_time
        - correlation_id
      properties:
        from_ip:
          type: string
          format: ip
          description: HTTP request remote IP address
          example: 192.168.1.1
        at_time:
          type: string
          format: date-time
          description: Timestamp of the action
          example: "2024-01-15T10:00:00Z"
        by_user:
          type: string
          description: User ID who performed the action
          pattern: '^[^\s]{1,40}$'
          example: user-123
        correlation_id:
          type: string
          description: Logging correlation ID
          pattern: '^[^\s]{1,40}$'
          example: corr-abc123
...
    Encounter:
      type: object
      description: Encounter domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
        - saved
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Encounter name
          pattern: '^[^\s]{1,40}$'
          example: my-Encounter
        description:
          type: string
          description: Encounter description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Encounter document
        status:
          type: string
          description: Encounter status
          enum:
            - active
            - archived
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'
        saved:
          $ref: '#/components/schemas/Breadcrumb'
    EncounterInput:
      type: object
      description: Input for creating a Encounter document
      properties:
        name:
          type: string
          description: Encounter name
          pattern: '^[^\s]{1,40}$'
          example: my-Encounter
        description:
          type: string
          description: Encounter description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Encounter document
        status:
          type: string
          description: Encounter status
          enum:
            - active
            - archived
          example: active
    EncounterUpdate:
      type: object
      description: Fields to update in a Encounter document (cannot include _id, created, saved)
      properties:
        name:
          type: string
          description: Encounter name
          pattern: '^[^\s]{1,40}$'
        description:
          type: string
          description: Encounter description
          pattern: '^[^\t\n]{0,255}$'
        status:
          type: string
          description: Encounter status
          enum:
            - active
            - archived
    Event:
      type: object
      description: Event domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Event name
          pattern: '^[^\s]{1,40}$'
          example: my-Event
        description:
          type: string
          description: Event description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Event document
        status:
          type: string
          description: Event status
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'
    EventInput:
      type: object
      description: Input for creating a Event document
      properties:
        name:
          type: string
          description: Event name
          pattern: '^[^\s]{1,40}$'
          example: my-Event
        description:
          type: string
          description: Event description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Event document
        status:
          type: string
          description: Event status
          example: active
...